py-geopolars/build.requirements.txt
# Dependencies
# Minimum version with binary dtype
polars>=0.14.19
pyarrow>=4.0
numpy>=1.16.0
# First version with latest naturalearth sample datasets for tests
geopandas>=0.12
shapely>=2.0

# Tooling
maturin==0.14.2
pytest==7.1.3
pytest-cov==3.0.0
hypothesis==6.54.5
black==22.8.0
blackdoc==0.3.6
isort==5.10.1
mypy==0.981
flake8==5.0.4
flake8-bugbear==22.8.23
flake8-comprehensions==3.10.0
flake8-docstrings==1.6.0
flake8-simplify==0.19.3
flake8-tidy-imports==4.8.0
pyupgrade==2.37.3

# Stub files
pandas-stubs==1.2.0.61
